requirements.txt
attrs==19.3.0
cliff==2.17.0
cmd2==0.9.22
cogapp==3.0.0
colorama==0.4.3
coverage==5.0
nose==1.3.7
pbr==5.4.4
prettytable==0.7.2
protobuf==3.11.1
pyaml==19.12.0
pyparsing==2.4.5
pyperclip==1.7.0
pyreadline==2.1
PyYAML==5.2
pyzmq==18.1.1
simplejson==3.17.0
six==1.13.0
stevedore==1.31.0
wcwidth==0.1.7
-e .